+++ zarrs/src/array/codec/array_to_bytes/packbits/packbits_codec.rs
@@ -3,9 +3,8 @@
 use std::{borrow::Cow, sync::Arc};
 
 use num::Integer;
-use zarrs_metadata::{
-    codec::packbits::PackBitsPaddingEncoding, v3::MetadataConfiguration, Endianness,
-};
+use zarrs_metadata::{v3::MetadataConfiguration, Endianness};
+use zarrs_metadata_ext::codec::packbits::PackBitsPaddingEncoding;
 use zarrs_plugin::PluginCreateError;
 use zarrs_registry::codec::PACKBITS;