+++ zarrs/src/array/codec/array_to_bytes/vlen_array.rs
@@ -24,7 +24,7 @@
 
 use crate::array::codec::array_to_bytes::vlen_v2::vlen_v2_macros;
 
-pub use crate::metadata::codec::vlen_array::{
+pub use zarrs_metadata_ext::codec::vlen_array::{
     VlenArrayCodecConfiguration, VlenArrayCodecConfigurationV1,
 };