+++ zarrs/src/array/codec/array_to_bytes/vlen_bytes.rs
@@ -20,7 +20,7 @@
 
 use crate::array::codec::array_to_bytes::vlen_v2::vlen_v2_macros;
 
-pub use crate::metadata::codec::vlen_bytes::{
+pub use zarrs_metadata_ext::codec::vlen_bytes::{
     VlenBytesCodecConfiguration, VlenBytesCodecConfigurationV1,
 };