+++ zarrs_metadata/src/v3/array/data_type.rs
@@ -1,5 +1,0 @@
-//! Zarr V3 data type metadata.
-//!
-//! See <https://zarr-specs.readthedocs.io/en/latest/v3/core/index.html#array-metadata-data-type>.
-//!
-//! This module is currently empty becaue no data types have any `configuration`.