+++ zarrs_metadata_ext/src/chunk_grid.rs
@@ -1,0 +1,6 @@
+//! Zarr array chunk grid metadata.
+//!
+//! See <https://zarr-specs.readthedocs.io/en/latest/v3/core/index.html#chunk-grid>.
+
+pub mod rectangular;
+pub mod regular;