+++ zarrs_data_type/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Bump `zarrs_metadata` to 0.5.0
+
 ## [0.2.0] - 2025-05-03
 
 ### Added